+++ packages/moonstone/Scroller/Scrollable.js
@@ -679,8 +679,13 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 				if (!this.childRef.containerRef.contains(spotItem)) {
 					return;
 				}
+
 				const
-					containerId = Spotlight.getActiveContainer(),
+					containerId = (
+						this.childRef.containerRef.dataset.containerId ||
+						this.childRef.contentRef.dataset.containerId ||
+						Spotlight.getActiveContainer()
+					),
 					direction = this.getPageDirection(keyCode),
 					rDirection = reverseDirections[direction],
 					viewportBounds = this.containerRef.getBoundingClientRect(),
@@ -694,7 +699,7 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 					this.animateOnFocus = false;
 					Spotlight.focus(next);
 				} else {
-					const nextPage = this.childRef.scrollToNextPage({direction, reverseDirection: rDirection, focusedItem: spotItem});
+					const nextPage = this.childRef.scrollToNextPage({direction, reverseDirection: rDirection, focusedItem: spotItem, containerId});
 
 					if (typeof nextPage === 'object') {
 						this.animateOnFocus = false;