+++ packages/moonstone/Scroller/Scroller.js
@@ -199,10 +199,10 @@ class ScrollerBase extends Component {
 		return node.getBoundingClientRect();
 	}
 
-	scrollToNextPage = ({direction, reverseDirection, focusedItem}) => {
+	scrollToNextPage = ({direction, reverseDirection, focusedItem, containerId}) => {
 		const
 			endPoint = this.getNextEndPoint(direction, focusedItem.getBoundingClientRect()),
-			next = getTargetByDirectionFromPosition(reverseDirection, endPoint, Spotlight.getActiveContainer());
+			next = getTargetByDirectionFromPosition(reverseDirection, endPoint, containerId);
 
 		return (next !== focusedItem) && next;
 	}